requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP client for API calls
httpx==0.25.2

# Environment variable management
python-dotenv==1.0.0

# Data validation
pydantic==2.5.0

# Airtable integration
pyairtable==2.2.3

# OpenAI integration
openai==1.3.7

# Telegram bot integration
python-telegram-bot==20.7

# Excel export functionality
openpyxl==3.1.2
pandas==2.1.4

# PDF generation
reportlab==4.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development and utilities
python-multipart==0.0.6 